move impl logic to user-impl module and switch doc language to english
