+++ services/user-service/user-impl/src/main/java/com/technokratos/eateasy/userimpl/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.technokratos.eateasy.userapi.config;
+package com.technokratos.eateasy.userimpl.config;
 
 import io.swagger.v3.oas.models.ExternalDocumentation;
 import io.swagger.v3.oas.models.OpenAPI;
