+++ services/user-service/user-impl/src/main/java/com/technokratos/eateasy/userimpl/exception/UserAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.technokratos.eateasy.userapi.exception;
+package com.technokratos.eateasy.userimpl.exception;
 
 public class UserAlreadyExistsException extends RuntimeException {
     public UserAlreadyExistsException(String message) {
