+++ services/user-service/user-impl/src/main/java/com/technokratos/eateasy/userimpl/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.technokratos.eateasy.userapi.exception;
+package com.technokratos.eateasy.userimpl.exception;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(String message) {
